Move tensorflow layer definitions to a new file
